+++ changelog.md
@@ -1,5 +1,5 @@
 # SG-Blocks Beta 1.6
-## th September 2025
+## th February 2026
 
 ***NEWS***
 - Added a new block "Sound Block"
@@ -8,8 +8,13 @@
 - Added a new mode in Input Block "Boolean", in other words "Yes or No" choice
 - Added a new example "Gambling Game" that using Probality Block and "Yes or No" choice
 - Added a new example "Music" that using Sound Block and Delay Block
+- Added a documentation
 
 ***CHANGES***
 - Set cursor pointer on hover select input
+- Now you can't connect a normal line from Conditon Block - Only True / False lines
+- Operation block displays the variable to which the result is written.
 
 ***FIXES***
+- Fixed the bug when you connect true/false and you key up the line become normal
+- Some UI fixes